Move "tab" related components into new folder
